added test case and fix for CAMEL-97
git-svn-id: https://svn.eu.apache.org/repos/asf/activemq/camel/trunk@565361 13f79535-47bb-0310-9956-ffa450edef68

+++ camel-core/src/main/java/org/apache/camel/processor/RedeliveryPolicy.java
@@ -113,7 +113,7 @@ public class RedeliveryPolicy implements Cloneable, Serializable {
     /**
      * Enables collision avoidence and sets the percentage used
      */
-    public RedeliveryPolicy collisionAvoidancePercent(short collisionAvoidancePercent) {
+    public RedeliveryPolicy collisionAvoidancePercent(double collisionAvoidancePercent) {
         useCollisionAvoidance();
         setCollisionAvoidancePercent(collisionAvoidancePercent);
         return this;
@@ -141,7 +141,7 @@ public class RedeliveryPolicy implements Cloneable, Serializable {
      * Sets the percentage used for collision avoidence if enabled via
      * {@link #setUseCollisionAvoidance(boolean)}
      */
-    public void setCollisionAvoidancePercent(short collisionAvoidancePercent) {
+    public void setCollisionAvoidancePercent(double collisionAvoidancePercent) {
         this.collisionAvoidanceFactor = collisionAvoidancePercent * 0.01d;
     }
